FEAT #76 Commit 4: extract BaseTurnDriver with shared decide/result/safety-cap helpers
Add BaseTurnDriver to agents/turn_driver.py with _make_result, _emit_safety_cap_callback,
and _try_decide (decide + per-run ValueError retry counter). StandardTurnDriver and
PlanningTurnDriver both inherit from it. PlanningTurnDriver drops its own _make_result,
_emit_safety_cap_callback, and the _plan_phase_failures instance dict; both plan and
reflect phases now delegate to _try_decide. Move consecutive_validation_failures from
PlanState onto AgentRun so the counter survives crashes and is scoped to the run, not
the plan revision.
