+++ src/agent_framework/agents/agent_run.py
@@ -35,5 +35,6 @@ class AgentRun:
     memory_projection_requests: tuple[str, ...] = ()
     in_parallel_batch: bool = False
     plan_state: "PlanState | None" = None
+    consecutive_validation_failures: int = 0
 
 __all__ = ["AgentRun"]
