+++ src/agent_framework/planning/plan_state.py
@@ -107,7 +107,6 @@ class PlanState:
     total_steps_executed: int = 0
     pending_callback_step_id: str | None = None
     awaiting_caller_callback: bool = False
-    consecutive_validation_failures: int = 0
 
 
 def plan_step_to_dict(step: PlanStep) -> dict[str, Any]:
